+++ src/redturtle/volto/adapters/stringinterp.py
@@ -4,6 +4,7 @@ from Products.CMFCore.interfaces import IContentish
 from redturtle.volto import _
 from zope.component import adapter
 
+
 try:
     from plone.stringinterp import _ as stringinterp_mf
 except ImportError: