+++ src/redturtle/volto/config/__init__.py
@@ -1,0 +1,4 @@
+import os
+
+
+MAX_LIMIT = int(os.environ.get("REDTURTLE_VOLTO_MAX_LIMIT_SEARCH") or 500)